autominer: an autoswitching miner for linux
autominer is a utility for cryptocurrency miners that selects the miner
and algorithm to mine based on past performance and current market rates
in order to achieve the optimal mining strategy.

Nicehash is the only hashpower market supported at this time.

Several miners are supported, see the miners directory for the list.
Miner applications are expected to be reachable via $PATH.

The config file is located at ~/.autominer/config, or
~/.autominer/$profile/config when run with --profile

To use autominer, the user runs nicehash-aggregator, which continuously
queries the nicehash api, computes aggregate pricing statistics, and
stores them on disk at $datadir/rates.

Next, the user runs autominer-mine, which executes the miner wrapper
scripts from the miners directory in order to determine which miners are
available, the algorithms they support, and to actually run the miner
application. The wrapper script monitors the miner in real time, and
writes performance statistics to stdout, which are read by
autominer-mine.

To see the aggregate pricing statistics, the user runs autominer-rates.

To see historical performance and miner/algorithm selections, the user
runs autominer-stats.

Signed-off-by: Todd Freed <todd.freed@gmail.com>