history: history and streams implementation
A stream is a sequence of mining intervals. Over the course of each
interval, a particular mining option is pursued. The interval contains
metadata about the option being pursued, e.g. its predicted and actual
profit, and hashrate observed from the miner. Each stream is identified
by a set of key/value pairs.

A history consists of a primary stream that includes a seequence of
consecutive intervals, and one or more secondary streams which include
some subset of the intervals which match some set of criteria.

autominer uses the history abstraction to record its selections, and
mining results, from the perspective of several streams. autominer stats
can be used subsequently to inspect the history and results on a
per-stream basis.

Signed-off-by: Todd Freed <todd.freed@gmail.com>